requirements.txt
fastapi
pydantic
sqlalchemy
alembic
uvicorn
asyncpg
greenlet
python-decouple
bcrypt
passlib
python-dotenv